Move value to runtime
