+++ core/src/runtime/mod.rs
@@ -1,3 +1,7 @@
 mod eval;
+mod value;
 
 pub(crate) use eval::{eval_expr, eval_query};
+pub(crate) use value::Show;
+
+pub use value::Value;
